Fix wrong behavior when changing pin code in settings
In security setting, if user change orientation during entering
pin code, the change will not apply successfully, it is because
the entered pin codes has not been saved and restored when
application was re-create after changed orientation. So we save
entered pin codes and restore them in onCreate when application
is re-created.

Change-Id: Ia8a3fb3ee75185ce028104487377e47f272905ce